# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Learning Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course with full hierarchy (units, chapters, modules)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course with hierarchy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── LOCATIONS ─────────────────────────────────────────────────────────────────
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List all centres/locations
      parameters:
        - name: zoneId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a new centre/location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Location created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"

  # ── PROGRESS / RATINGS ───────────────────────────────────────────────────────
...
  /dashboard/trainer/{id}:
    get:
      operationId: getTrainerDashboard
      tags: [dashboard]
      summary: Trainer-specific dashboard summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trainer dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainerDashboard"
  /dashboard/coordinator/{id}:
    get:
      operationId: getCoordinatorDashboard
      tags: [dashboard]
      summary: Coordinator-specific dashboard summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coordinator dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoordinatorDashboard"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # AUTH
...
    TrainerStats:
      type: object
      required: [trainerId, name, studentCount, completionRate]
      properties:
        trainerId:
          type: integer
        name:
          type: string
        studentCount:
          type: integer
        completionRate:
          type: number

    # STUDENT
...
    Course:
      type: object
      required: [id, name, syllabus]
      properties:
        id:
          type: integer
        name:
          type: string
        syllabus:
          type: string
    CourseInput:
      type: object
      required: [name, syllabus]
      properties:
        name:
          type: string
        syllabus:
          type: string
    CourseDetail:
      type: object
      required: [id, name, syllabus, units]
      properties:
        id:
          type: integer
        name:
          type: string
        syllabus:
          type: string
        units:
          type: array
          items:
            $ref: "#/components/schemas/Unit"
    Unit:
      type: object
      required: [id, name, orderIndex, chapters]
      properties:
        id:
          type: integer
        name:
          type: string
        orderIndex:
          type: integer
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
    Chapter:
      type: object
      required: [id, name, orderIndex, modules]
      properties:
        id:
          type: integer
        name:
          type: string
        orderIndex:
          type: integer
        modules:
          type: array
          items:
            $ref: "#/components/schemas/Module"
    Module:
      type: object
      required: [id, name, orderIndex]
      properties:
        id:
          type: integer
        name:
          type: string
        orderIndex:
          type: integer

    # LOCATION
    Location:
      type: object
      required: [id, centreName, zoneId, zoneName]
      properties:
        id:
          type: integer
        centreName:
          type: string
        zoneId:
          type: integer
        zoneName:
          type: string
        coordinatorId:
          type: ["integer", "null"]
        coordinatorName:
          type: ["string", "null"]
    LocationInput:
      type: object
      required: [centreName, zoneId]
      properties:
        centreName:
          type: string
        zoneId:
          type: integer
        coordinatorId:
          type: ["integer", "null"]

    # PROGRESS / RATINGS
    ProgressRecord:
      type: object
      required: [id, studentId, studentName, moduleId, moduleName, chapterName, unitName, rating, submittedAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        trainerId:
          type: ["integer", "null"]
        trainerName:
          type: ["string", "null"]
        moduleId:
          type: integer
        moduleName:
          type: string
        chapterName:
          type: string
        unitName:
          type: string
        rating:
          type: string
          enum: [totally_dependent, physical_prompting, verbal_prompting, clueing, independent]
        comments:
          type: ["string", "null"]
          description: Optional trainer comments (max 500 chars)
        submittedAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    TrainerDashboard:
      type: object
      required: [trainerId, trainerName, zoneName, studentCount, completionRate, ratingDistribution, recentActivity]
      properties:
        trainerId:
          type: integer
        trainerName:
          type: string
        zoneName:
          type: string
        studentCount:
          type: integer
        completionRate:
          type: number
        ratingDistribution:
          type: array
          items:
            $ref: "#/components/schemas/RatingCount"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ProgressRecord"
    CoordinatorDashboard:
      type: object
      required: [coordinatorId, coordinatorName, zoneName, centreCount, trainerCount, studentCount, completionRate, trainerPerformance]
      properties:
        coordinatorId:
          type: integer
        coordinatorName:
          type: string
        zoneName:
          type: string
        centreCount:
          type: integer
        trainerCount:
          type: integer
        studentCount:
          type: integer
        completionRate:
          type: number
        trainerPerformance:
          type: array
          items:
            $ref: "#/components/schemas/TrainerStats"
    RatingCount:
      type: object
      required: [rating, count]
      properties:
        rating:
          type: string
        count:
          type: integer

    # ACCOUNT CREATION
...